docs(license): Use MIT license to cover project
The `adm` coding project uses the MIT license as its core license for all
source code and documentation.

This project abides by the REUSE specification version 3.0[1]. As per section
`License Files` of the REUSE specification, the MIT license exists within the
`LICENSES/` directory as `MIT.txt`[2].

The `adm` coding project uses the REUSE specification to make it clear what
portions of the project is under a given license, and who owns the copyright to
that portion of the project.

[1]: https://reuse.software/spec/
[2]: https://reuse.software/spec/#license-files

Signed-off-by: Jason Pena <jasonpena@awkless.com>