feat(adm::vcs): Define execute()
The execute() function is defined separately in order to decouple it from any
future VCS tool that the `adm` project plans to use (mainly Git). The function
also makes it easier to test whether the command-line VCS tool is executing
properly, or I somehow messed up the call to the VSC command-line tool.

Signed-off-by: Jason Pena <jasonpena@awkless.com>